+++ klink_plugin/python/klink_server/methods/__init__.py
@@ -13,6 +13,7 @@ from . import pcell_m  # noqa: F401
 from . import selection_m  # noqa: F401
 from . import edit_m  # noqa: F401
 from . import view_m  # noqa: F401
+from . import d25_m  # noqa: F401
 from . import events_m  # noqa: F401
 from . import exec_m  # noqa: F401
 from . import drc_m  # noqa: F401